requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
lighthouse==0.1.0
